Move mocks into their own files.
