+++ SharedPackages/VPN/Tests/VPNTests/Mocks/MockDNSResolver.swift
@@ -1,0 +1,34 @@
+//
+//  MockDNSResolver.swift
+//
+//  Copyright © 2025 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import Foundation
+@testable import VPN
+
+final class MockDNSResolver: DNSResolving {
+    var receivedEndpoints: [Endpoint?]?
+    var results: [Result<Endpoint, DNSResolutionError>?]
+
+    init(results: [Result<Endpoint, DNSResolutionError>?]) {
+        self.results = results
+    }
+
+    func resolveSync(endpoints: [Endpoint?]) -> [Result<Endpoint, DNSResolutionError>?] {
+        receivedEndpoints = endpoints
+        return results
+    }
+}
