+++ SharedPackages/VPN/Tests/VPNTests/WireGuardAdapterTests.swift
@@ -460,93 +460,3 @@ final class WireGuardAdapterTests: XCTestCase {
 private enum TestError: Error, Equatable {
     case someError
 }
-
-// MARK: - Mocks
-
-private final class MockWireGuardInterface: WireGuardGoInterface {
-    var turnOnCallCount = 0
-    var lastTurnOnConfig: String?
-    var lastTurnOnHandle: Int32?
-    var lastTurnOnResult: Int32?
-    var turnOnReturnHandle: Int32 = 7
-
-    var turnOffCallCount = 0
-    var lastTurnOffHandle: Int32?
-
-    var setConfigCallCount = 0
-    var lastSetConfigHandle: Int32?
-    var lastSetConfig: String?
-    var setConfigResult: Int64 = 0
-
-    var bumpSocketsCallCount = 0
-    var disableRoamingCallCount = 0
-
-    var getConfigReturnValue: UnsafeMutablePointer<CChar>?
-
-    var loggerContext: UnsafeMutableRawPointer?
-    var loggerFunction: (@convention(c) (UnsafeMutableRawPointer?, Int32, UnsafePointer<CChar>?) -> Void)?
-
-    func turnOn(settings: UnsafePointer<CChar>, handle: Int32) -> Int32 {
-        turnOnCallCount += 1
-        lastTurnOnConfig = String(cString: settings)
-        lastTurnOnHandle = handle
-        let result = turnOnReturnHandle
-        lastTurnOnResult = result
-        return result
-    }
-
-    func turnOff(handle: Int32) {
-        turnOffCallCount += 1
-        lastTurnOffHandle = handle
-    }
-
-    func getConfig(handle: Int32) -> UnsafeMutablePointer<CChar>? {
-        getConfigReturnValue
-    }
-
-    func setConfig(handle: Int32, config: String) -> Int64 {
-        setConfigCallCount += 1
-        lastSetConfigHandle = handle
-        lastSetConfig = config
-        return setConfigResult
-    }
-
-    func bumpSockets(handle: Int32) {
-        bumpSocketsCallCount += 1
-    }
-
-    func disableSomeRoamingForBrokenMobileSemantics(handle: Int32) {
-        disableRoamingCallCount += 1
-    }
-
-    func setLogger(context: UnsafeMutableRawPointer?, logFunction: (@convention(c) (UnsafeMutableRawPointer?, Int32, UnsafePointer<CChar>?) -> Void)?) {
-        loggerContext = context
-        loggerFunction = logFunction
-    }
-}
-
-private final class MockDNSResolver: DNSResolving {
-    var receivedEndpoints: [Endpoint?]?
-    var results: [Result<Endpoint, DNSResolutionError>?]
-
-    init(results: [Result<Endpoint, DNSResolutionError>?]) {
-        self.results = results
-    }
-
-    func resolveSync(endpoints: [Endpoint?]) -> [Result<Endpoint, DNSResolutionError>?] {
-        receivedEndpoints = endpoints
-        return results
-    }
-}
-
-private final class MockTunnelFileDescriptorProvider: TunnelFileDescriptorProviding {
-    var fileDescriptor: Int32?
-
-    init(fileDescriptor: Int32?) {
-        self.fileDescriptor = fileDescriptor
-    }
-
-    func currentFileDescriptor() -> Int32? {
-        fileDescriptor
-    }
-}
